refactor: イベントハンドラをeventHandlers.tsに分離
